Generating QFileOpenEvent messages when S60 UI receives OpenFileL
S60 applications can receive OpenFileL requests from other apps in the
OpenFileL method of their main document class. This S60 main document
class is internal to QtGui, so the request is turned into a
QFileOpenEvent for the application to pick up if it want to implenment
this service.

If an application wants to receive the QFileOpenEvent before UI
construction, it should do something like:
    app->sendPostedEvents(app, QEvent::FileOpen);

Task-number: QTBUG-15015
Reviewed-by: Shane Kearns