QStyleSheetStyle: Fix crash that occurs with several instance of QStyleSheetStyle
The problem is that when a widget is destroyed, it was not properly removed
from the cache.  We connected the destroyed signal to a slot in the
QStyleSheetStyle for the first QStyleSheetStyle that handle a widget.
but if this QStyleSheetStyle is destroyed, that connection is lost.

Solution:  Create a new QStyleSheetStyleCaches that will be responsible
to clean the caches. This objects is not destroyed as long as there is
a QStyleSheetStyle instance, so the connection is not lost.

I took the oportunity to move all the caches to this object.

Reveiwed-by: Gabriel
Task-number: QTBUG-11658