Fix detection of N95 with linux 2.6.34+
Linux cdc-acm driver detects the debug port on some Nokia phones automatically
When it does, it creates two ttyACMx devices and the symlink in
/dev/serial/by-id contains the CDC master interface number.

In contrast the generic usbserial driver creates a symlink containing the
CDC slave interface number. This is because it knows nothing about CDC
and is just looking for BULK IN/OUT interfaces that might be serial ports.

This change creates match strings for both the master and slave interfaces,
so the phone is detected if either driver is loaded.

Reviewed-By: axis