Added a file handle to QFileOpenEvent on Symbian
QFileOpenEvent needs to tell the recipient what file to open. Normally
that is by filename. But on Symbian, some files can only be opened from
an already-open file handle that has been passed in. For example data
caged files in c:\private or z:\sys. So QFileOpenEvent needs to be
able to hold a Symbian file handle so that the recipient can access
any file requested of it.

Coming soon... a method by which the reciever can access the file
without knowing about all this messy file handle stuff going on behind
the scenes.

Task-number: QTBUG-15015
Reviewed-by: Shane Kearns