Prevent multiple modprobe messages for the same phone
Cleaned up some unnecessary code in the serenum_unix file.
Add the device only once to the list used for printing info about
devices with no serial drivers.
Add each ACM interface ID, trusting the numbers given in the master
interface. The existing code of searching for the slave interface
would have been appropriate for a user-mode device driver, but we only
need the interface numbers as we are going to use existing device drivers
via the tty devices

Reviewed-By: axis