added an openFile method to QFileOpenEvent
An application needs some way to get an open QFile when a QFileOpenEvent
contains a Symbian file handle. Rather than exposing that file handle
and forcing platform specific code to the application, this method
hides the details giving an effective universal method for turning a
QFileOpenEvent into an open QFile.

Task-number: QTBUG-15015
Reviewed-by: Shane Kearns