qdoc: Second pass at generating a map.
All the files are there now. It outputs each element as a <topicref>.
We might want to output the entries for C++ classes, namespaces,
and QML classes differently. But the validator is able to use this
map to validate everything, and there are no errors. So until we
get further instructions, this is it.

Task-number: QTBUG-14981