Don't crash if QScriptClass property getter returns an invalid value
It's possible that a class claims to have a property of a given name
(i.e. queryProperty() returns true), but returns an invalid value
for that property. In that case we should silently convert the value
to undefined, otherwise the value may wreak havoc in JS.

This is a regression from Qt 4.5 (pre-JavaScriptCore-based), which
had this check.

Task-number: QTBUG-15079
Reviewed-by: Jedrzej Nowacki