QS60Style: Color calculation should be optimized
Currently, QS60Style calculates some palette colors (tooltip base and
button). Since native side does not have a color for these, but a
nine-part theme graphic, the style tries to estimate the color of the
bitmap by taking a small sample of the QPixmap and calculate the
RGB colors of 32*32 pixels. This is rather slow, it takes around
110 msecs for each QApplication, when the application is started.
Note that color is cached to member variable of style, but it is very
rarely asked again (as the color is polished to all widgets/apps)
and the cache is not shared across processes.

As a fix, style now calculates the button color (tooltip color
is no longer calculated, as no other QStyle does that, and
tooltips do not anyway work in the Qt/Symbian) and stores the
calculated value to Global QSettings together with active theme
ID. Now, when a second Qt application is launched, the stored
theme ID value is matched with currently active theme. If it
matches, then the stored Button color is used. Otherwise, color
is again calculated and stored. If theme is unchanged, the
application launch is ~95msecs faster.

Task-number: QTBUG-14860
Reviewed-by: Jani Hautakangas