Send WinIdChange event when winId is set to zero
This allows an observer to get a notification just before the window
handle owned by a native widget is destroyed.  Note that, at the
point when the event is sent, the widget's internalWinId() will
return the new value, but the old native window handle will not
be destroyed until after the event handler is run.

Task-number: QTMOBILITY-645
Reviewed-by: sroedal