Fix a behaviour change of sizeHint() introduced by 6d4d265e7e67dde58
Commit 6d4d265e7e67dde58e45d7d89f4974d0bd8b70e4 added a behaviour change
in the cases if there was an item with height-for-width and sizeHint()
was called with no constraint.  The commit tried to return the height
needed for the preferred width, but it still did not satisfy the
constraints, since the width used as the constraint could be less than
the preferred width.  This also meant that the sizeHint(Qt::MinimumSize)
could actually be larger than the smallest possible size.

The behaviour should be that it should return the smallest width
possible regardless of height. For instance, for a label it could return
the size of the longest word (to avoid hyphenation issues).

The same logic applies for the height: It should return the smallest
height possible regardless of width. For instance, for a label it could
then return the height of the font.

I also had to fix some stuff in the heightForWidthWithSpanning()
autotest since it wrongly expected the maximum size to be
QWIDGETSIZE_MAX in several of the cases.  However, that is the current
behaviour (and it is a bug), but it is unrelated to the problem with
spans so I simply fix the test and mark them with QEXPECT_FAIL.

Reviewed-by: John Tapsell