ShortcutOverride has no effect on some shortcuts on Mac OS X (Cocoa).
When generating the native key sequences for menu items, Qt prefers the
non private unicode characters. But the characters in the NSEvent for
keyboard events can contain characters form the unicode range reserved
for Apple.
For e.g. when user presses the "delete" key, the event contains
NSDeleteFunctionKey, where in Qt is expecting NSDeleteCharacter. For now
this is the only key identified for translation. If we find similar
translations, those can be added to qt_mac_removePrivateUnicode().

Task-number: QTBUG-12495
Reviewed-by: Denis