SSL: Fix for systemCaCertificates being called first on symbian
On symbian, thread names must be unique (actually kernel object names)
When a thread exits, there may still be open handles, for example a
debugger or RUndertaker so the thread name cannot be reused immediately.

S60 has an RUndertaker instance in a background thread, which is used
to display the "application closed" messages when a crash happens.
Until that thread has run and checked the thread exit to see if it was
a crash or not, the thread remains open.

When systemCaCertificates is called as the first API call, it calls itself
via ensureinitialised() to set the default CA certs. This double call
should be addressed by QTBUG-15218.

In any case, QSslSocket::systemCaCertificates() is intended to refresh
from the system - if application code calls it too quickly in succession it
could also trigger this bug.

Task-number: QTBUG-15126
Reviewed-by: Markus Goetz