Add Location as self signable capability in patch_capabilities.pl
Location was not originally included in self-signable set, since in
S60 3.1 devices it is not so. However, it is a lesser of two evils to
make applications with Location capability not install on 3.1 devices
than make them not work correctly on any devices, so Location is now
included in the self-signable set in patch_capabilities.pl.

Task-number: QTBUG-13891
Reviewed-by: Janne Koskinen