Switched qdesktopservices to use SchemeHandler for Symbian^3 and later.
SchemeHandler is plugin extensible component to handle several
different URI schemes. By default schemes such as http, https, rtsp,
mailto, file, wtai, tel and cti are supported.
In some devices/platforms also other schemes might be supported.

This commit only defines the USE_SCHEMEHANDLER macro and re-orders
the qdesktopservices_s60 source code to make it more readable and to
make it compile both with and without the flag. The actual
implementation was provided at the same time when qdesktopservices
for symbian was initially implemented.

Why support is enabled only for Symbian^3 and later? SchemeHandler
component does not exist in S60 3.1 - 5.0 public SDKs, actually it does
not exist also in Forum Nokia provided Symbian^3 SDK, but is available
in Symbian^3 PDK. Since building Qt for Symbian^3 anyway require PDK
and because the release builds for Symbian^3 are done separately
from older Symbian versions, PDK dependency should be Ok.

It might be that SchemeHandler headers will be relicensed as part of Qt,
then SchemeHandler support also for older S60/Symbian releases can be
enabled. If this happens it will be separate commit.

Task-number: QTBUG-15282
Reviewed-by: Miikka Heikkinen