Fix possible missing glyphs in text when using GL engine
If you create/destroy gl contexts a lot, you may sometimes get a new
context with the same pointer as a destroyed context. When you look up
the glyph cache in the font engine using the context pointer as a key,
you will then get a glyph cache which contains no valid data. We need
to reset the glyph cache completely in this case and set up bindings for
the new context so that the glyph cache can be repopulated and reused.

Note that there is a different solution for this in Qt 4.8, so this is
temporary solution for the Qt 4.7.x series.

Task-number: QT-4162
Reviewed-by: Fabien Freling