Have qmake generate V8 builtins
Using the js2c.py script. This adds a dependency on Python, but
makes our lives easier when updating src/3rdparty/v8. It also
demonstrates that we _can_ use qmake rather than SCons to fully
build V8 without having to check in generated files, if we want
to go that route. (The snapshot feature is still missing, though.)