QNetworkAccessManager: enable synchronous HTTP calls
To enable synchronous calls, an attribute in the QNetworkRequest has
to be set. If set, when QNetworkAccessManager::get() (and post(),
put()) returns, the reply is finished and all data has been read in
case of success.
This feature is semi-public for now (usable, but not documented).
To enable this, an attribute in the QNetworkRequest must be set. If
this attribute is set, we open a new connection to the server with
only one channel and call the channels' sockets' waitFor* methods.

Reviewed-by: Markus Goetz