Don't destroy Qt's internal pthread_key_t if it was not initialized
The destructor function destroy_current_thread_data_key() was always
called on library unload, even if the pthread_key_t was not initialized.
This can cause us to destroy a key that doesn't belong to us, as pointed
out in the task below.

Fix this by creating a function static instance of a class whose
destructor will destroy the key.

Task-number: QTBUG-10861
Reviewed-by: Morten Sørvig