Cocoa: combobox does not hightlight when mouse re-hovers the first item
The reason is that on Mac, the highlight is supposed to switch off when
the mouse moves out of the drop down menu. On X11, it stays. So there is
a separate code path for this in qcombobox.cpp. But it fails to clear the
index set in the view when the mouse leaves, which stops the item from
re-highligh when the mouse re-enters.

Reviewed-by: ogoffart