Fix QPixmap::fromImage() in the OpenVG pixmap backend.
If QPixmap::fromImage() was called with a QImage that has an alpha
channel and with the 'Qt::NoOpaqueDetection' image conversion flag set,
then we would always create an opaque (RGB32) pixmap. This is
incorrect.

The fix is to check if the source QImage has an alpha channel and if
so, use sourceFormat(), otherwise use RGB32.

Task-number: QT-4278
Reviewed-by: Jani Hautakangas