Prevent compilers optimizing eval timebomb code out of existence.
The variable that holds the eval license key is a placeholder that is
patched during package installation.  Unfortunately, for a non-final
package build, the placeholder is filled with nulls at compile-time and
a clever compiler will optimize away most of the eval timebomb code due
to a check in the eval code for the first character of the license key
being null.

This commit makes the variable that holds the license key volatile, to
convince compilers that they cannot make assumptions about the contents
of the variable when optimizing.

Task-number: QT-3848
Acked-by: Thiago Macieira