Smoke test for QAbstractFileEngine
This tests the default file support provided by Qt, together with
QFSFileEngine, QResource engine and a reference custom file engine that
works on top of a simple virtual in-memory file system.

For now, the test is only focusing on QFile I/O functionality, it should
be extended to cover QFileInfo, QDir, QDirIterator and file system
operations in QFile.

The intent for the reference file engine is not to be a fully functional
file system, but to ensure that minimal support for custom file engines
is working.

Reviewed-by: Shane Kearns