Fix possible artifacts under glyphs in texture glyph cache
We would disregard the first glyph in each line when calculating the
required height of the line in the glyph cache. If the first glyph
was taller than any of the other glyphs in the same line, the glyph
drawn underneath it in the cache could potentially overlap it, and
you would see it as dots or lines underneath the glyph in the output.

Task-number: QTBUG-14806
Reviewed-by: Jiang Jiang