Fix "array subscript out of bounds" warning with GCC 4.5 in QtOpenGL
If you have:
    qreal m[2][2];
    const qreal *constData() { return m[0]; }

And somewhere else:
    const qreal *data = matrix.constData();
    for (int i = 0; i < 2 * 2; ++i)
        tmp = data[i];

Then GCC "sees through" the data pointer to know that the type is
actually a qreal[2] and will complain when you access data[2] and
data[3] in the example above.

It seems to be satisfied with this return *m, even though it's exactly
the same thing. This is probably fragile.