Fix strict-alias breaking warnings with GCC.
GCC doesn't like any kind of reinterpret_cast or C-style cast with
pointers. So instead do the work with static_cast<>, which it seems to
like.

Also took the opportunity to change the generic payload type to void*,
so the alignment works as expected. I wonder how we haven't had
serious crashes so far on ARM...

Reviewed-by: Samuel Rødal