QKqueueFileSystemWatcherEngine: Unlock mutex between two events.
In the worker thread unlock the mutex between processing two events.
Otherwise it's possible for the worker thread to block the application
thread when many events occur. Also, there's no need to lock the mutex
when processing a pipe event. Generally the worker thread should hamper
the application thread as little as possible, so only lock the mutex
where needed.

Merge-request: 2425
Reviewed-by: João Abecasis <joao.abecasis@nokia.com>