QKqueueFileSystemWatcherEngine: Use higher file descriptors.
A file descriptor is used for every path to be monitored, but descriptors
below FD_SETSIZE (typically 1024) are precious, for use with select(2). To
allow the application (and other parts of Qt) to use select(2), try to
duplicate the descriptor returned by open(2) above FD_SETSIZE and close(2)
the original. However, only do so when the descriptor table is already
fairly large (FD_SETSIZE / 2). This keeps the descriptor table small for
applications that use only a few descriptors.

While here, also set the close-on-exec flag on the (new) descriptor.

Merge-request: 2425
Reviewed-by: João Abecasis <joao.abecasis@nokia.com>