QPollingFileSystemWatcherEngine: Fix double report of directory change.
The polling engine first retrieves a QFileInfo for a given path, then
tests whether it's different from before and if so, stores the new file
info and emits a signal. In case path is a directory the test also
checks if the list of directory entries has changed. This creates a
window between retrieving the file info and the test in which a file can
be added/removed from the directory or the directory itself can be
removed. In that case the test returns true, because the list of entries
has changed, but outdated file info is stored which means that on the
next timeout the same change will be reported a second time.

Therefore, refresh the file info after the test for changes.

Merge-request: 2425
Reviewed-by: João Abecasis <joao.abecasis@nokia.com>