QKqueueFileSystemWatcher: don't stop thread that isn't running
When removing paths from the watch list, if we end up with an empty
watch list, we would send a request to the processing thread to quit.

In the case where the watch list was empty to begin with (the paths
being removed weren't actually being watched) the request to quit would
still be queued. If the processing thread wasn't running (and it
shouldn't if there weren't any paths being watched) the request to quit
would still be posted but not processed.

The next time paths were added and the thread started, the request would
be processed and the thread would quit at once.

When removing paths from the list, we now check whether the watch list
is empty to begin with and exit early without asking the processing
thread to quit itself.

Task-Number: QTBUG-14435
Reviewed-by: Bradley T. Hughes