Fix buffered/unbuffered mode issues on symbian
Due to a bug in the symbian file server, files in /resource can't be
opened for unbuffered read, only for default mode read. (it doesn't mask
the cache control flags when doing the security check)
So read will always be done in default mode.
Symptom of this was that QML plugin loading failed as the plugin description
in /resource could not be read.

Buffered or unbuffered writes (i.e. whether the cache should be write
through or write behind) are controlled by the QIODevice::Unbuffered
flag, therefore it needs to be passed through to the file engine.

An optimisation for unix and windows to force unbuffered mode in the file
engine is inappropriate (as that is referring to buffering in the standard
library, which is bypassed entirely by using the low level RFile to open
files on symbian)

Reviewed-by: joao