Add symbian platform security and case sensitivity test cases
On symbian, check that opening files in protected locations either
succeeds or fails cleanly as expected by the capabilities the test is
compiled with.
This acts as a regression test for opening files in /resource, and also
checks the other locations behave as expected.

On all platforms, check file created with a mixed case filename
can be accessed by it's aliases that differ only in case (or not)
depending on the OS case sensitivity.
For Windows/Symbian, file should be accessible by altered case aliases
For other OS, it should not (other OS supported by Qt are unix like)

Reviewed-by: mread