Cleaned up sis_targets.prf
Removed the unnecessary checks for the case where pkg file doesn't exist
are no longer needed as we now have dependencies to pkg files and no
sis creation targets are generated if pkg is not generated.

Also improved documentation a bit and added the missing
ok_installer_sis target generation.

Reviewed-by: Janne Koskinen