purge vestiges of imageformat-plugins
The imageformat-plugins variable is never referenced by the .pr[io] for
building Qt. Its presence in configure.exe appears to be a vestige of
an earlier state of affairs.

As a reminder, jpeg, mng, tiff and gif can be built into QtGui or built
as plugins. The default is plugin. During configuration, this can be
overridden by adding, for example, "no-jpeg" to disable support
altogether or "jpeg" to build into QtGui.

Merge-request: 961
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@nokia.com>