Fix text disappearing on GL when RGB-path is taken (no transformation)
When there is no transformation on a glyph in the GL texture glyph
cache, it will take the RGB code path to support subpixel antialiasing.
In this code path we would ask glTextImage2D() to convert from a
GL_ALPHA buffer to an internal format of GL_RGBA. This is not supported
on OpenGL ES 2, and would cause missing text for some drivers.

The change also fixes a typo in an #ifdef.

Task-number: QTBUG-15789
Reviewed-by: Kim