Fix QScriptEngine in threads
v8::Initialize should not be called when we use isolate, this function
is not thread safe and initialize a "default" isolate if we do not plan
to use threads.

But removing it show cases where we do not enter the Isolate properly.
That is, there was a QScriptIsolate missing in ~QScriptValue