Helper functions for synchronously waiting on an active object to complete
QtNetwork APIs are based on calling a function asychronously.
However the user may call a waitForXXXX function later to convert this
into a synchronous call.
For Symbian, the async call should be implemented as an active object.
These helpers make the synchronous conversion possible.

Unix used select() in a polling way to do async mode and blocking way
for the synchronous conversion.

Reviewed-by: mread