Ensure that DBus is connected before all uses.
In early system startup applications may try to use ICD before it is
contactable. Ensure that the connection to ICD is established before all
calls are made.

If ICD is still not contactable QDBusServiceWatcher is used to monitor
registration of the com.nokia.icd2 address and reconnection is attempted
once ICD is started.

Task-number: Maemo 199755