Fix QWingedEdge memory usage issue
QWingedEdge should reserve the amount of segments/edges/vertices in
relative to the element count instead of length of the QPainterPath
passed in. This problem is especially severe when QTextLayout using
it to calculate the region for full line selection highlighting,
because the region used QFIXED_MAX as right most coordinate, it will
cost more than 2 GB of memory allocated just for it (recorded by
valgrind on Mac OS X), and cause crash in systems short of memory.

Task-number: QTBUG-15823
Reviewed-by: Samuel Rødal