Remove Lighthouse specific code in QGLWidget constructor
truth to be told, I can't remember why I decided that it was a good idea
to have the code there. But clearly now I don't think so anymore. Such
initialisation should happen in qgl_qpa.cpp chooseContext.

If we tear down the window and the context then that has to be solved
elsewhere.