Fix a rare race condition when showing windows
When setting a window to translucent shortly before showing it with the
OpenGL graphics system, the hijacking would call XDestroyWindow without
resetting the internal state. This might lead to the window waiting for
a map notify message forever.

The patch now properly resets the internal state when destroying the old
window, so it'll get mapped correctly when show() is called.

Reviewed-by: Denis Dzyubenko