Missing glyphs in GL when scaling QStaticText and QML text items
Since scaled text is backed by an A8 glyph cache, it needs to repopulate
even if it has previously populated the cache for the unscaled text.
This means that we need to record in the userData that the type of the
cache is not the same as the last time. Otherwise the A8 version of the
cache will only be updated when it's created and when the text actually
changes, leading to glyphs missing on screen when zooming text using
a scale.

Task-number: QTBUG-16012
Reviewed-by: Gunnar