Improving memory leak analysis for argv, envp
QtMainWrapper() gets the argv and envp arrays from __crt0(). There have
been memory leaks reported from QtMainWrapper() because the contents of
argv and envp are not deleted when the function returns. There has been
some attempt to delete the memory with "delete []" on the top level
arrays, but this will not delete any memory allocated for the strings
themselves.

The big problem is that there is no function which will free argv and
envp properly. QtMainWrapper() can only guess at what is the best way
to free the arrays. If QtMainWrapper() reverses the implementation of
__crt0(), it is then tied to the current implementation of that
function in an unrelated component.

Instead we want these arrays to not be viewed as memory leaks when the
app exits. QrMainWrapper() is only ever called as a top level function
before main(), and the memory will be released automatically when the
app exists shortly after the funciton returns. So in practice there is
no meaningful leak from not freeing these arrays.

The memory leak analysis tools should be able to spot heap ownership
rooted in static data, and not treat that as a leak. In this case it
is valid to put the argv and envp pointers in static data. Doing this
should stop leaks being reported from these arrays.

Task-number: QTBUG-15687
Reviewed-by: axis