QDateTimeEdit is not showing correctly when resizes in symbian
The problem boils down to using large square area for QDateTime widget
(which is more or less a date-aware combobox). Style tries to make the
combobox button as wide as it is tall, and as button is as tall as
the combobox frame, it steals the whole square widget for itself.

To fix this, we now limit the combobox button height to be no more
than fourth of the total width of the widget.
Thus, in normal, non-resized case, button is, as tall as the frame,
but as widget is made larger, button stops to grow after reaching
quarter of width of the widget.

Task-number: QT-4312
Reviewed-by: Janne Koskinen