Fix qwidget test crash
CEikonEnv::Static()->AppUiFactory()->Cba() call always returns NULL
unless cba was defined via resources, which we are no longer doing.

Due to the buggy implementation of MEikAppUiFactory interface in
Symbian, the only way to get non-resources created cba is to use the
return value of CEikonEnv::Static()->AppUiFactory()->SwapButtonGroup()
method, so changed the test cases to do a dummy swap. Since the swap
just changes the value of a member variable and doesn't trigger
anything, it should not cause any interference for the tests.

Task-number: QTBUG-15915
Reviewed-by: Janne Koskinen
Reviewed-by: Sami Merila