Fixed a bug in the input methods on S60 with QGraphicsWebView.
The bug happened because we did not expect to receive a call to
CancelFepInlineEdit() when the widget had not previously been given
any preedit text by S60. However, this can happen, and if it happens
at the same time that there is a temporarily visible character in the
widget (because it will be obscured as a password a second later),
then the temporary character would be discarded.

Fixed by checking in the beginning of CancelFepInlineEdit() whether
we currently have any temporary preedit text or not, and if we do, we
don't clear the contents, since it is the input context itself that
generated the preedit text, not S60.

Also made sure that the temporary preedit text is committed if we
receive a call to UpdateFepInlineTextL(), which may sometimes happen
without a call to StartFepInlineEditL().

Task:     QT-4314
AutoTest: Included
RevBy:    Shane Kearns