Skip failing tests
For Mac OS X 10.5 and earlier, QDir::canonicalPath depends on
QDir::cleanPath which is fundamentally broken and specifically fails for
the case of "/./", returning "" instead of "/".

Fixing cleanPath to a sane and correct behaviour seems to break more
code than it is worth so we're skipping these individual tests for the
time being.