Improve performance of clipping to a scaled QRectF
Clipping to a scaled QRectF was so far hitting the slow
code path in the raster paintengine that generates clip spans
for every line. This is not needed as we also clip translated
QRectF's to integer rects.

This patch does the same for scaled rectangles and ensures the
clipping really happens at the closest pixel boundary.

Reviewed-by: Samuel