Fixes text jitter in QtQuick components when using the virtual keyboard
QGraphicsView::mapFromScene() returns a QRect instead of a
QRectF. Thus, we were prematurely rounding the micro focus rectangle
coordinates in QGraphicsView::inputMethodQuery(). (The
Qt::ImMicroFocus query returns a QRectF from QGraphicsScene, which is
the result of querying QDeclarativeTextEdit and mapping it to the
scene.)

Auto-test included.

Reviewed-by: leo
Reviewed-by: bnilsen
Task-number: QTCOMPONENTS-288
Task-number: QTBUG-16063