doc: Replaced some \raw and \endraw uses with \table and \endtable
In DITA XML, there is no straightforward way to translate raw
html into DITA XML, because the XML stream writer automagically
escapes all the raw html elements. So I am beginning to replace
uses of the \raw command with \table, which gets output correctly.
The problem is the XML stream writer must see each XML element
start and end, because it keeps them on a stack. When you output
XML elements with the writeCharacters() function, it escapes the
'<' and '>' of any XML elements the character string contains.