Cocoa: popup hides behind window
Reason: when cocoa receives a mouse press/release in a window, it
finds the correct view inside that window and sends the mouse event
to it. But NSWindow also does some other stuff just before sending the
event, like raise and lower windows. So when we override sendEvent, and
more over, do not call [super sendEvent], we stop this raise/lower etc
from working. So, to make this work again, I partially revert change
0b2eab87ad3bd73a0744469a45c29ca098649c9b

Task-number: QTBUG-15638, QTBUG-1517
Reviewed-by: Fabien Freling